Update - Static Library for Assets
Assets were removed from the Game Faculties library in order to allow
modular use of one without the other.

The directory structure remains unchanged as the Assets use a file
defined within Game Faculties. (ie. "AbstractAsset.h")

Added a method overload for Font::Print() to provide an interface for
printing and specifying the location to print at simultaneously.

Recompiled when the work was done.