Update - Design-Hack
I added a Sprite::Load() overload to deal with the fact the original
Load() is tightly coupled with the GameFaculties. I hadn't considered
the possibility, during the design process, that an asset may be
manually loaded and that the GameFaculties might not be used whatsoever.

This hack allows sprites to be manually loaded by forcing the user to
provide a valid Texture* and GLSLProgram*. It does not fix the root
design flaw; consequently a user may use the first Load() and receive
errors due to prerequisite assets not being available.

THIS IS A HACK NOT A FIX!