Update - Manual Load - Build
Implemented what I think will be a suitable procedure for manually
loading assets. It forces the assets to be recorded. Of course the user
may still allocate an asset manually and attempt to load a file, which
may potentially crash their application.

The only solution I can think of would be to add exception catching code
to the Load() methods of the asset types. Alternatively GetAsset() could
force a load when the Asset can't be found. This may be ideal for the
majority of situations.