Control Module Attachment
The control module should always be the first GameModule attached to the
Game::Instance().

What I did was ensure that is enforced. It doesn't truly matter at this
point in time as it would always be processed before anything else in
any of my projects. When or if I, in the future, create other modules
which overload the Process() method it can and will be important to
maintain this order.