Logger Integration
I was unsure of how to integrate the logging tools in a way they were
meant to be used. So I went a bit lower level than what was intended.

I decided the log should be managed by the project itself rather than by
the LogFile_Manager singleton. This choice made integration much easier.
Though the other way is possible with the changes this choice brought
about.

So Asset_Faculties create a logger::Log and a logger::FilePolicy. It
thus creates its very own Log File to manage. This Log File is provided
to all the separate management objects through the Asset_Faculties
singleton.