Doxygen Output
I haven't uploaded this documentation yet so I figured I would do that
now, before it get even more out of date.

It takes like ~30 minutes to process in doxygen, so I will update it
when I feel like it or otherwise get around to it. As such, this is out
of date already as previously mentioned.