render bugfix: Function AnchorBottomLeft(...)
Bug: Sprites would not render.

Summary: I was calculating the xyz vertices of the sprites to face away
from the camera, pretty sure this wasn't the only problem with the
calculations either.

Bottom line is the shaders were fine after all, as I had thought they
should be in the beginning. This given they worked in every other
project in which they were used. I had thought the shaders were the
culprit because the Triangle Render test wouldn't work when the shaders
used any matrix other than the model matrix. The cause for THIS is still
unknown, it doesn't matter much now though.