Implemented Input Design
Input Polling with function callbacks for actions/ranges.

This design allows multiple devices to be implemented and polled
together (asynchronously). The currently only the steam controller is
integrated with the design. Next will be keyboard, and then xbox
controllers. The polling is implemented such that input states are
checked, and if the correct conditions are met a callback is made to
whichever function is bound to that particular input condition.