compiler if guards - bitset union
MSVC shits the bed if you try to compile a bitset in a union.
You have to put the bitset inside a struct and union the struct.

GCC is the opposite. It shits the bed if you put the bitset in an
anonymous struct.
It is perfectly fine with a bitset in a union otherwise though.

Added if guards to select and compile the correct code for MSVC.