StackTracer - Added New Library Project (dll/lib)
Extracted StackTracer from the Logger project, and re-included as an
independent library. StackTracer is compiled as both a Dynamic Library,
and a Static Library.

The dynamic library exports a single function which prints a stack trace
to a log.  Sorry there is only the one, this is my first dll binary.

The static library does it all.